requirements.txt
# 核心依赖
requests>=2.28.0
python-dotenv>=0.19.0

# 类型提示支持 (Python < 3.8)
typing-extensions>=4.0.0

# 文档处理依赖
pdfplumber>=0.9.0      # PDF文件处理
python-docx>=0.8.11    # DOCX文件处理
markdown>=3.4.0        # Markdown文件处理

# 向量数据库依赖
pymilvus>=2.4.0        # Milvus向量数据库客户端

# LLM依赖
google-genai>=0.9.0    # Google Gemini API官方SDK
Pillow>=9.0.0          # 图像处理支持

# 可选依赖 - 用于更好的开发体验
# pytest>=7.0.0  # 用于单元测试
# black>=22.0.0   # 代码格式化
# flake8>=4.0.0   # 代码检查
